실행시켰을때 원하는 페이지로 수정

+++ CRUDboard/src/main/webapp/WEB-INF/views/board/view.jsp
@@ -10,9 +10,6 @@
 <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.bundle.min.js"></script>
 
 <style>
-.navbar-custom {
-	background-color: #343a40;
-} /* 네비게이션 바 색상을 Bootstrap의 기본 어두운 회색으로 변경 */
 .btn-custom {
 	border-color: #007bff;
 	color: #007bff;
@@ -21,9 +18,7 @@
 	background-color: #007bff;
 	color: white;
 } /* 버튼을 호버했을 때의 배경과 글씨 색상을 설정 */
-body {
-	font-size: 1.2rem;
-} /* 본문의 글씨 크기를 1.2rem으로 설정 */
+
 </style>
 </head>
 <body>  
@@ -49,7 +44,7 @@ body {
 	
 	<div>
 		<a href="/board/modify?bno=${view.bno}" class="btn btn-outline-primary">게시물 수정</a>
-		<a href="/board/delete?bno=${view.bno}" class="btn btn-outline-danger">게시물 삭제</a>
+		<a href="/board/delete?bno=${view.bno}" class="btn btn-outline-danger" onclick="return confirm('정말 삭제하시겠습니까?')">게시물 삭제</a>
 	</div>
 </div>
